--- a/spec/support/database_cleaner.rb
+++ b/spec/support/database_cleaner.rb
@@ -1,7 +1,7 @@
 RSpec.configure do |config|
 
   config.before(:suite) do
-    DatabaseCleaner.strategy = :truncation, { pre_count: true }
+    DatabaseCleaner.strategy = :truncation, { pre_count: true, reset_ids: false }
     DatabaseCleaner.clean_with(:truncation)
   end