@marcgreenstock 2015-03-09
This is the new speed in Travis:

```
Finished in 39.52 seconds (files took 6 seconds to load)
493 examples, 0 failures, 8 pending
```